requirements.txt
flask
numpy
joblib
scikit-learn
python-dotenv
joblib
numpy
groq